… tests/fault_tolerance/deploy/README.md
… tests/fault_tolerance/deploy/test_deployment.py
… tests/fault_tolerance/deploy/worker_config.py