#### Overview:

Using ai-perf to generate loads for k8 FT tests

#### Details:

Using ai-perf to generate loads for k8 FT tests

**Recovery time Calculation**
Eliminated timezone mixing: All timestamps now come from container logs (UTC).
Simplified logic: No longer need to track fault_time from test.log.txt
More accurate: Directly measures actual container restart time
Before (Incorrect):
Test Log (Local): 2025-09-30T12:00:00 - Container killed
Container Log (UTC): 2025-09-30T19:00:30 - Container restarted
Calculated Recovery: 25230 seconds ❌ (includes 7-hour timezone difference)
After (Correct):
Previous Container Log (UTC): 2025-09-30T19:00:00 - Last log before termination
New Container Log (UTC): 2025-09-30T19:00:30 - First log after restart  
Calculated Recovery: 30 seconds ✅ (actual recovery time)

Example output
```
============================================================
╒═══════════════════════════════════╤════════════════════════════════════════════════════╕
│ Metric                            │ Value                                              │
╞═══════════════════════════════════╪════════════════════════════════════════════════════╡
│ Test Directory                    │ test_fault_scenario[sglang-agg-tp-1-dp-1-frontend] │
├───────────────────────────────────┼────────────────────────────────────────────────────┤
│ Number of Clients                 │ 10                                                 │
├───────────────────────────────────┼────────────────────────────────────────────────────┤
│                                   │                                                    │
├───────────────────────────────────┼────────────────────────────────────────────────────┤
│ === Deployment Metrics ===        │                                                    │
├───────────────────────────────────┼────────────────────────────────────────────────────┤
│ Startup Time                      │ 69.00 sec                                          │
├───────────────────────────────────┼────────────────────────────────────────────────────┤
│ Recovery Time                     │ 1.00 sec                                           │
├───────────────────────────────────┼────────────────────────────────────────────────────┤
│                                   │                                                    │
├───────────────────────────────────┼────────────────────────────────────────────────────┤
│ === Request Metrics ===           │                                                    │
├───────────────────────────────────┼────────────────────────────────────────────────────┤
│ Total Requests                    │ 1318                                               │
├───────────────────────────────────┼────────────────────────────────────────────────────┤
│ Successful Requests               │ 1302                                               │
├───────────────────────────────────┼────────────────────────────────────────────────────┤
│ Failed Requests                   │ 16                                                 │
├───────────────────────────────────┼────────────────────────────────────────────────────┤
│ Success Rate                      │ 98.79%                                             │
├───────────────────────────────────┼────────────────────────────────────────────────────┤
│                                   │                                                    │
├───────────────────────────────────┼────────────────────────────────────────────────────┤
│ === Latency Metrics (seconds) === │                                                    │
├───────────────────────────────────┼────────────────────────────────────────────────────┤
│ Mean Latency                      │ 0.502                                              │
├───────────────────────────────────┼────────────────────────────────────────────────────┤
│ P50 Latency                       │ 0.391                                              │
├───────────────────────────────────┼────────────────────────────────────────────────────┤
│ P90 Latency                       │ 0.415                                              │
├───────────────────────────────────┼────────────────────────────────────────────────────┤
│ P99 Latency                       │ 0.761                                              │
├───────────────────────────────────┼────────────────────────────────────────────────────┤
│                                   │                                                    │
├───────────────────────────────────┼────────────────────────────────────────────────────┤
│ === Token Generation Metrics ===  │                                                    │
├───────────────────────────────────┼────────────────────────────────────────────────────┤
│ Time to First Token (mean)        │ 0.205 sec                                          │
├───────────────────────────────────┼────────────────────────────────────────────────────┤
│ Inter-Token Latency (mean)        │ 0.0031 sec                                         │
├───────────────────────────────────┼────────────────────────────────────────────────────┤
│                                   │                                                    │
├───────────────────────────────────┼────────────────────────────────────────────────────┤
│ === Throughput Metrics ===        │                                                    │
├───────────────────────────────────┼────────────────────────────────────────────────────┤
│ Total Throughput                  │ 19.72 req/s                                        │
├───────────────────────────────────┼────────────────────────────────────────────────────┤
│ Avg Client Throughput             │ 1.97 req/s                                         │
╘═══════════════════════════════════╧════════════════════════════════════════════════════╛
```


#### Where should the reviewer start?

tests/fault_tolerance/deploy/client.py: Core ai-perf client implementation with retry logic and streaming support
tests/fault_tolerance/deploy/parse_results.py: Updated parsing logic for AI-Perf JSON outputs and recovery time calculation
tests/fault_tolerance/deploy/README.md: Documentation of the new testing architecture and metrics

#### Related Issues: (use one of the action keywords Closes / Fixes / Resolves / Relates to)

DIS-679

<!-- This is an auto-generated comment: release notes by coderabbit.ai -->

## Summary by CodeRabbit

- New Features
  - Switched fault-tolerance load generation to AI-Perf with configurable tokens, concurrency, retries, streaming, and warmup.
  - Per-client artifact directories and summarized metrics (success/fail counts, latency p50/p90/p99, TTFT, inter-token, throughput).
  - Multi-test processing and combined summary output with optional SLA checks.

- Documentation
  - README rewritten to reflect AI-Perf usage, updated commands, outputs, and file references.

- Tests
  - Increased default client retry attempts to 3 and introduced retry_delay (replacing max_request_rate).
  - Enhanced log parsing and recovery-time calculation; new CLI supports single or multiple test directories.

<!-- end of auto-generated comment: release notes by coderabbit.ai -->